--- new-mock-engine-store/src/utils.rs
@@ -1,27 +1,0 @@
-// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
-
-pub use std::{
-    cell::RefCell,
-    collections::{BTreeMap, HashMap, HashSet},
-    pin::Pin,
-    sync::{
-        atomic::{AtomicU64, Ordering},
-        Mutex,
-    },
-    time::Duration,
-};
-
-pub use engine_store_ffi::{
-    interfaces::root::DB as ffi_interfaces, EngineStoreServerHelper, RaftStoreProxyFFIHelper,
-    RawCppPtr, UnwrapExternCFunc,
-};
-pub use engine_traits::{
-    Engines, Iterable, KvEngine, Mutable, Peekable, RaftEngine, RaftLogBatch, SyncMutable,
-    WriteBatch, CF_DEFAULT, CF_LOCK, CF_RAFT, CF_WRITE,
-};
-pub use kvproto::{
-    raft_cmdpb::AdminCmdType,
-    raft_serverpb::{RaftApplyState, RaftLocalState, RegionLocalState},
-};
-pub use protobuf::Message;
-pub use tikv_util::{box_err, box_try, debug, error, info, warn};
